+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * `StripPrefix (prefix :: Symbol)`: You can remove a certain prefix from your field names
   * `CamelTo (separator :: Symbol)` and its variants, `CamelToSnake` and `CamelToKebab`: Transform field names written
   in CamelCase to snake\_case, kebab-case, or with a custom separator.
+* Remove redundant metadata about the query nature when logging the query
 ## 0.0.1.0 -- 2021-11-05
 
 * First version. Released on an unsuspecting world.